+++ app/controllers/poll/participation.js
@@ -1,5 +1,5 @@
 import Controller, { inject as controller } from '@ember/controller';
-import User from '../../native-models/user';
+import User from '../../models/user';
 import { inject as service } from '@ember/service';
 import { action } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
