+++ app/models/poll.js
